+++ redis-pro/Views/RedisEditorView/SetEditorView.swift
@@ -117,16 +117,14 @@ struct SetEditorView: View {
     
     
     func onDeleteConfirmAction(_ index:Int) -> Void {
-        globalContext.alertVisible = true
-        globalContext.showSecondButton = true
-        globalContext.primaryButtonText = "Delete"
-        
         let item = list[index] ?? "''"
-        globalContext.alertTitle = String(format: Helps.DELETE_LIST_ITEM_CONFIRM_TITLE, item)
-        globalContext.alertMessage = String(format:Helps.DELETE_LIST_ITEM_CONFIRM_MESSAGE, item)
-        globalContext.primaryAction = {
-            try deleteEle(index)
-        }
+        
+        globalContext.confirm(String(format: Helps.DELETE_LIST_ITEM_CONFIRM_TITLE, item)
+                              , alertMessage: String(format:Helps.DELETE_LIST_ITEM_CONFIRM_MESSAGE, item)
+                              , primaryAction: {
+                                try deleteEle(index)
+                              }
+                              , primaryButton: "Delete")
         
     }
